chore: rename to sigsegv_handler
