+++ bin/reth/src/lib.rs
@@ -151,10 +151,10 @@ pub mod rpc {
 }
 
 #[cfg(all(unix, any(target_env = "gnu", target_os = "macos")))]
-pub mod so_handler;
+pub mod sigsegv_handler;
 
 #[cfg(not(all(unix, any(target_env = "gnu", target_os = "macos"))))]
-pub mod so_handler {
+pub mod sigsegv_handler {
     //! Signal handler to extract a backtrace from stack overflow.
     //!
     //! This is a no-op because this platform doesn't support our signal handler's requirements.
