+++ bin/reth/src/main.rs
@@ -13,7 +13,7 @@ fn main() {
     use reth::cli::Cli;
     use reth_node_ethereum::EthereumNode;
 
-    reth::so_handler::install();
+    reth::sigsegv_handler::install();
 
     // Enable backtraces unless a RUST_BACKTRACE value has already been explicitly provided.
     if std::env::var("RUST_BACKTRACE").is_err() {
